Fixes regression issue with order of state truncation vs commit (#362)
It's important that the commit happens before state truncation. This regressed with the merge of the Batching system, it's unknown why. There is now test coverage for the specific behaviour. The effect should be zero duplicates upon healthy rebalancing.

+++ parallel-consumer-core/src/test/java/io/confluent/parallelconsumer/truth/CommitHistorySubject.java
@@ -41,7 +41,7 @@ public class CommitHistorySubject extends Subject {
         return assertAbout(commitHistories()).that(actual);
     }
 
-    public void atLeastOffset(int needleCommit) {
+    public void atLeastOffset(long needleCommit) {
         Optional<Long> highestCommitOpt = this.actual.highestCommit();
         check("highestCommit()").about(OptionalSubject.optionals())
                 .that(highestCommitOpt)
